Implement vacuum items feature in BrewKettleBlockEntity
The commit added a feature in BrewKettleBlockEntity that enables it to vacuum items from its surroundings. The code tries to find item entities in a defined area around the block, and if there is enough space in its container, it absorbs the item entities. The item entities are then removed from the world. The changes also include related helper and utility methods to facilitate this new feature.