Updated heat sources and recipe handling, and bump version to 9.2.0
This commit includes changes to heat sources in Growthcraft, adding bonfire and brazier from decorative blocks. It also refactors the Mixing Vat recipes, removing unnecessary category variable and simplifying recipe parsing. Finally, the version of the mod is bumped to 9.2.0, which is reflected in the mod's metadata, README, and gradle properties.